requirements.txt
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
py-healthcheck==1.10.1
six==1.16.0
Werkzeug==2.2.3
